+++ src/dryad/format/endf/test/CMakeLists.txt
@@ -3,6 +3,6 @@ add_cpp_test( format.endf.createBoundaries   createBoundaries.test.cpp )
 add_cpp_test( format.endf.createInterpolant  createInterpolant.test.cpp )
 add_cpp_test( format.endf.createInterpolants createInterpolants.test.cpp )
 
-add_cpp_test( format.endf.tabulatedCrossSection tabulatedCrossSection.test.cpp )
-add_cpp_test( format.endf.reaction              reaction.test.cpp )
-add_cpp_test( format.endf.projectileTarget      projectileTarget.test.cpp )
+add_cpp_test( format.endf.createTabulatedCrossSection createTabulatedCrossSection.test.cpp )
+add_cpp_test( format.endf.reaction                    reaction.test.cpp )
+add_cpp_test( format.endf.projectileTarget            projectileTarget.test.cpp )
